+++ .changeset/afraid-trainers-occur.md
@@ -1,0 +1,5 @@
+---
+'svelte': patch
+---
+
+fix: ensure use directives execute in the correct sequence
